split entry update / delete logic out from record helpers
